Add kernel by utility function and minor changes to testbench and FFT

+++ include/sw-ops/FFT.hpp
@@ -9,7 +9,7 @@ namespace sw {
 
 const double PI = 3.141592653589793238460;
 
-typedef std::complex<double> Complex;
+typedef std::complex<float> Complex;
 typedef std::valarray<Complex> CArray;
 
 void fft(CArray &x) {
