+++ tests/sw-ops/convFFT.cpp
@@ -5,12 +5,13 @@
 
 const double PI = 3.141592653589793238460;
 
-typedef std::complex<double> Complex;
+typedef std::complex<float> Complex;
 typedef std::valarray<Complex> CArray;
 
 int main() {
   const Complex test[] = {1.0, 1.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0};
   CArray data(test, 8);
+  
 
   // forward fft
   ama::sw::fft(data);
